refactor(server): Factor out metadata factory function
